+++ src/components/BarChart/types.ts
@@ -51,6 +51,7 @@ export interface YAxisOptions {
   labelFormatter: NumberLabelFormatter;
   labelColor: string;
   backgroundColor: string;
+  integersOnly: boolean;
 }
 
 export interface Annotation {