+++ src/components/LineChart/LineChart.tsx
@@ -135,6 +135,7 @@ export function LineChart({
     labelFormatter: (value: number) => value.toString(),
     labelColor: DEFAULT_GREY_LABEL,
     backgroundColor: 'transparent',
+    integersOnly: false,
     ...yAxisOptions,
   };