+++ src/components/MultiSeriesBarChart/tests/Chart.test.tsx
@@ -77,6 +77,7 @@ describe('Chart />', () => {
       labelFormatter: (value: number) => value.toString(),
       labelColor: 'red',
       backgroundColor: 'transparent',
+      integersOnly: false,
     },
     gridOptions: {
       showHorizontalLines: true,