+++ src/components/MultiSeriesBarChart/types.ts
@@ -66,4 +66,5 @@ export interface YAxisOptions {
   labelFormatter: NumberLabelFormatter;
   labelColor: string;
   backgroundColor: string;
+  integersOnly: boolean;
 }